Update amp.hbs
amp-youtube script added, as according to  issue #39075 .  
If i want to add features like autopause, that will come to play If the user has interacted with the YouTube video (e.g., mutes/unmutes, pauses/resumes, etc.), and the video is scrolled in or out of view, the state of the video remains as how the user left it. , where should i add its box? i got a bit confused that's why I didn't added it.